Add Lowry Delphic Phenomenon

+++ api/internal/domain/candleStick.go
@@ -1,8 +1,8 @@
 package domain
 
 type CandleStick struct {
-	CloseTime    float64
-	OpenTime     float64
+	CloseTime    int64
+	OpenTime     int64
 	Close        float64
 	Open         float64
 	Max          float64
